Moved event fabric into separate file
